[add] Add public & gp_message
